Add image handling to elements

+++ src/src/Model/Table/ElementsTable.php
@@ -151,10 +151,12 @@ class ElementsTable extends Table
      */
     public function afterSave(Event $event, EntityInterface $entity, ArrayObject $options): void
     {
-        $previousImage = WWW_ROOT . 'img' . DS . 'element-img' . DS . $options['previousImageName'];
+        if (isset($options['previousImageName'])) {
+            $previousImage = WWW_ROOT . 'img' . DS . 'element-img' . DS . $options['previousImageName'];
 
-        if (is_file($previousImage)) {
-            unlink($previousImage);
+            if (is_file($previousImage)) {
+                unlink($previousImage);
+            }
         }
     }
 }
